remove a homework post

+++ backend/src/api/v1/contacts.ts
@@ -35,17 +35,25 @@ app.post('/add', async (req, res) => {
 })
 
 app.post('/remove', async (req, res) => {
-    Contact.deleteOne({
-        firstName: req.body.firstName,
-        lastName: req.body.lastName,
-        phone: req.body.phone,
-    }, undefined, (err: any) => {
-        if (err) {
-            res.send({ title: "error", message: err.message });
-        } else {
-            res.send({ message: "deleted" });
-        }
-    });
+    const userInfo = await User.findOne({ sessionId: req.body.sessionId})
+    if(userInfo == null){
+        res.status(406);
+        res.json({ title: "user doesn't exist"});
+    }
+    else{
+        Contact.deleteOne({
+            user : userInfo._id,
+            firstName: req.body.firstName,
+            lastName: req.body.lastName,
+            phone: req.body.phone,
+        }, undefined, (err: any) => {
+            if (err) {
+                res.send({ title: "error", message: err.message });
+            } else {
+                res.send({ message: "deleted" });
+            }
+        });
+    }
 })
 
 export default app;
